Add files via upload
Repository for a machine learning project focused on diagnosing breast cancer using a RandomForest classifier. This project includes code for data loading, preprocessing, model training with hyperparameter tuning via GridSearchCV, and evaluation metrics such as accuracy, recall, precision, and specificity. To demonstrate the use of RandomForest in healthcare analytics, we created this project as a part of an academic assignment. Our focus was on utilizing classification techniques to identify early signs of breast cancer.